pyproject.toml
[project]
name = "my-vllm"
version = "0.1.0"
description = "vLLM deployment for myself"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "python-dotenv>=1.1.0",
    "lmcache>=0.3.0",
    "vllm>=0.9.1",
    "torch>=2.7.0",
]

[tool.uv]
index-url = "https://mirror.nju.edu.cn/pypi/web/simple"

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
    { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "black>=25.1.0",
]
